+++ tests/django_test_project/settings.py
@@ -2,8 +2,12 @@
 
 SECRET_KEY = "django-bizcal-tests"
 INSTALLED_APPS = [
+    "django.contrib.admin",
     "django.contrib.auth",
     "django.contrib.contenttypes",
+    "django.contrib.messages",
+    "django.contrib.sessions",
+    "django.contrib.staticfiles",
     "django_bizcal",
 ]
 DATABASES = {
@@ -15,4 +19,21 @@ DATABASES = {
 USE_TZ = True
 TIME_ZONE = "UTC"
 ROOT_URLCONF = "tests.django_test_project.urls"
-
+MIDDLEWARE = [
+    "django.contrib.sessions.middleware.SessionMiddleware",
+    "django.contrib.auth.middleware.AuthenticationMiddleware",
+    "django.contrib.messages.middleware.MessageMiddleware",
+]
+TEMPLATES = [
+    {
+        "BACKEND": "django.template.backends.django.DjangoTemplates",
+        "APP_DIRS": True,
+        "OPTIONS": {
+            "context_processors": [
+                "django.contrib.auth.context_processors.auth",
+                "django.contrib.messages.context_processors.messages",
+            ]
+        },
+    }
+]
+STATIC_URL = "/static/"